requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.4.0
Django==1.11.20
django-cors-headers==2.5.2
djangorestframework==3.9.2
enum34==1.1.6
ipython==5.8.0
ipython-genutils==0.2.0
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
